feat(base): encode AVP list cardinality at the type level
Adds Has*AVPs list mixin interfaces alongside the existing singular
Has*AVP mixins, so callers know from the type whether an AVP may
appear multiple times. Audits all RFC 6733 messages against the
spec grammar and corrects cardinality, missing AVPs, and wrong
placement (Destination-Host/Realm on Request only).

+++ sparta-diameter-base/src/main/java/com/sipgate/sparta/diameter/base/session/DiameterNodeConfig.java
@@ -173,8 +173,8 @@ public final class DiameterNodeConfig {
      * The application capabilities advertised in CER/CEA.
      */
     public record Capabilities(
-            List<Integer> authApplicationIds,
-            List<Integer> acctApplicationIds,
+            List<Long> authApplicationIds,
+            List<Long> acctApplicationIds,
             List<Long> supportedVendorIds) {
 
         public Capabilities {
